# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/operators": {
      "get": {
        "description": "Returns one row per operator, joined to the latest month available in `agg_monthly_by_operator`. Includes BOE share from Vaca Muerta wells.",
        "operationId": "OperatorsController_list_v1",
        "parameters": [
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Sort key for the latest-month totals.",
            "schema": {
              "default": "boe",
              "type": "string",
              "enum": [
                "oil_m3",
                "gas_thousand_m3",
                "boe",
                "active_wells"
              ]
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/OperatorListItemDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          }
        },
        "summary": "List operators with latest-month totals",
        "tags": [
          "operators"
        ]
      }
    },
    "/api/v1/operators/{slug}": {
      "get": {
        "description": "Metadata, latest-month rank, latest-month totals and YTD totals for one operator.",
        "operationId": "OperatorsController_detail_v1",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "example": "ypf",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/OperatorDetailDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Operator slug not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Operator detail",
        "tags": [
          "operators"
        ]
      }
    },
    "/api/v1/operators/{slug}/production": {
      "get": {
        "description": "One point per `date_month`. Optionally filtered with `from` / `to`.",
        "operationId": "OperatorsController_production_v1",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "example": "ypf",
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Inclusive lower bound (YYYY-MM or YYYY-MM-DD).",
            "schema": {
              "example": "2026-01",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Inclusive upper bound (YYYY-MM or YYYY-MM-DD).",
            "schema": {
              "example": "2026-04",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/OperatorTimeSeriesPointDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Operator slug not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Monthly time series for an operator",
        "tags": [
          "operators"
        ]
      }
    },
...
    "/api/v1/prices/commodities/{commodity}": {
      "get": {
        "description": "Monthly value history for one commodity (`gold`, `silver`, `copper`, `lithium_etf`, `uranium_etf`).",
        "operationId": "PricesController_commoditySeries_v1",
        "parameters": [
          {
            "name": "commodity",
            "required": true,
            "in": "path",
            "schema": {
              "example": "gold",
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Inclusive start (YYYY-MM or YYYY-MM-DD).",
            "schema": {
              "example": "2025-06",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-05",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CommoditySeriesPointDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Series not tracked.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Commodity time series",
        "tags": [
          "prices"
        ]
      }
    },
...
    "/api/v1/prices/fuel": {
      "get": {
        "description": "Filterable list of station-level fuel prices (current snapshot). Use `provincia`, `producto`, `bandera`, `from`, `to`.",
        "operationId": "PricesController_fuel_v1",
        "parameters": [
          {
            "name": "provincia",
            "required": false,
            "in": "query",
            "description": "Province name (case-insensitive exact match).",
            "schema": {
              "example": "NEUQUEN",
              "type": "string"
            }
          },
          {
            "name": "producto",
            "required": false,
            "in": "query",
            "description": "Product label (case-insensitive exact match).",
            "schema": {
              "example": "Nafta Premium",
              "type": "string"
            }
          },
          {
            "name": "bandera",
            "required": false,
            "in": "query",
            "description": "Bandera (brand) — case-insensitive exact match.",
            "schema": {
              "example": "YPF",
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Inclusive lower bound on fecha_vigencia.",
            "schema": {
              "example": "2026-05-01",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-05-31",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 1000,
              "default": 100,
              "example": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/FuelPriceRowDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/PaginationMetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta",
                    "pagination"
                  ]
                }
              }
            }
          }
        },
        "summary": "Gas station fuel prices",
        "tags": [
          "prices"
        ]
      }
    },
    "/api/v1/prices/fuel/latest": {
      "get": {
        "description": "For each (provincia, producto) pair, returns the latest fecha_vigencia plus min / avg / max across stations on that date. Optionally filter by `provincia` and / or `producto`.",
        "operationId": "PricesController_fuelLatest_v1",
        "parameters": [
          {
            "name": "provincia",
            "required": false,
            "in": "query",
            "schema": {
              "example": "NEUQUEN",
              "type": "string"
            }
          },
          {
            "name": "producto",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Nafta Premium",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/FuelLatestEntryDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                }
              }
            }
          }
        },
        "summary": "Latest fuel prices by province + product",
        "tags": [
          "prices"
        ]
      }
    },
...
  },
  "info": {
    "title": "Petroldata.ar API",
    "description": "REST API over Argentina Secretaría de Energía oil & gas production data (focus: Vaca Muerta).\n\n## Response envelope\n\nEvery non-GeoJSON endpoint returns:\n\n```json\n{\n  \"data\": ...,\n  \"meta\": {\n    \"source\": \"Secretaría de Energía / datos.energia.gob.ar\",\n    \"dataset\": \"Producción de petróleo y gas por pozo\",\n    \"license\": \"CC-BY-4.0\",\n    \"last_source_update\": \"2026-04-01\",\n    \"last_ingested_at\": \"2026-05-27T14:30:00.000Z\",\n    \"vaca_muerta_filter\": \"formation + unconventional + sub-tipo\"\n  },\n  \"pagination\": { \"page\": 1, \"limit\": 50, \"total\": 1000 }\n}\n```\n\n`pagination` is only present on paginated list endpoints. GeoJSON endpoints (`/geo/*`) return raw GeoJSON.\n\n## Errors\n\n```json\n{ \"error\": { \"code\": \"NOT_FOUND\", \"message\": \"Well not found: 99999\" } }\n```\n\n## Units\n\n- `oil_m3` — cubic metres of oil\n- `oil_bbl` — barrels of oil (1 m³ ≈ 6.2898 bbl)\n- `oil_bbl_d` — barrels per day (oil_bbl ÷ days in month)\n- `gas_thousand_m3` — thousands of cubic metres of gas (Mm³)\n- `gas_mcf` — thousand cubic feet of gas\n- `gas_mmcf_d` — million cubic feet per day\n- `gas_mm3_d` — million m³ per day\n- `boe` — barrels of oil equivalent (oil_bbl + gas_boe)\n\n## Vaca Muerta flag\n\n`vm_combined = formation_vaca_muerta AND unconventional AND sub_resource_type ∈ {Shale, Shale Oil, …}` — pre-computed in the Python pipeline.",
...
  "servers": [
...
  "components": {
    "schemas": {
...
      "MetaDto": {
        "type": "object",
        "properties": {
          "source": {
            "type": "string",
            "example": "Secretaría de Energía / datos.energia.gob.ar"
          },
          "dataset": {
            "type": "string",
            "example": "Producción de petróleo y gas por pozo"
          },
          "license": {
            "type": "string",
            "example": "CC-BY-4.0"
          },
          "last_source_update": {
            "type": "object",
            "example": "2026-04-01",
            "nullable": true
          },
          "last_ingested_at": {
            "type": "string",
            "example": "2026-05-27T14:30:00.000Z"
          },
          "vaca_muerta_filter": {
            "type": "string",
            "example": "formation + unconventional + sub-tipo"
          }
        },
        "required": [
          "source",
          "dataset",
          "license",
          "last_source_update",
          "last_ingested_at",
          "vaca_muerta_filter"
        ]
      },
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 50
          },
          "total": {
            "type": "number",
            "example": 1000
          }
        },
        "required": [
          "page",
          "limit",
          "total"
        ]
      },
...
      "OperatorListItemDto": {
        "type": "object",
        "properties": {
          "operator_slug": {
            "type": "string",
            "example": "ypf"
          },
          "operator_name": {
            "type": "string",
            "example": "YPF S.A."
          },
          "latest_month": {
            "type": "string",
            "example": "2026-04-01"
          },
          "oil_m3": {
            "type": "number",
            "example": 1848981.5
          },
          "oil_bbl_d": {
            "type": "number",
            "example": 387657.95
          },
          "gas_thousand_m3": {
            "type": "number",
            "example": 888132.1
          },
          "gas_mmcf_d": {
            "type": "number",
            "example": 1045.49
          },
          "boe": {
            "type": "number",
            "example": 44029598.96
          },
          "active_wells": {
            "type": "number",
            "example": 15290
          },
          "vm_share_boe": {
            "type": "number",
            "example": 0.811,
            "description": "Share of total BOE in the latest month coming from Vaca Muerta wells."
          }
        },
        "required": [
          "operator_slug",
          "operator_name",
          "latest_month",
          "oil_m3",
          "oil_bbl_d",
          "gas_thousand_m3",
          "gas_mmcf_d",
          "boe",
          "active_wells",
          "vm_share_boe"
        ]
      },
      "OperatorLatestDto": {
        "type": "object",
        "properties": {
          "oil_m3": {
            "type": "number",
            "example": 1757724.93
          },
          "oil_bbl_d": {
            "type": "number",
            "example": 368525.06
          },
          "gas_thousand_m3": {
            "type": "number",
            "example": 870517.38
          },
          "gas_mmcf_d": {
            "type": "number",
            "example": 1024.74
          },
          "boe": {
            "type": "number",
            "example": 42813011.69
          },
          "active_wells": {
            "type": "number",
            "example": 5044
          }
        },
        "required": [
          "oil_m3",
          "oil_bbl_d",
          "gas_thousand_m3",
          "gas_mmcf_d",
          "boe",
          "active_wells"
        ]
      },
      "OperatorYtdDto": {
        "type": "object",
        "properties": {
          "year": {
            "type": "number",
            "example": 2026
          },
          "oil_m3": {
            "type": "number",
            "example": 7217666.24
          },
          "oil_bbl": {
            "type": "number",
            "example": 45397749.4
          },
          "gas_thousand_m3": {
            "type": "number",
            "example": 3520367.61
          },
          "gas_mcf": {
            "type": "number",
            "example": 124320726.67
          },
          "boe": {
            "type": "number",
            "example": 173823921.59
          }
        },
        "required": [
          "year",
          "oil_m3",
          "oil_bbl",
          "gas_thousand_m3",
          "gas_mcf",
          "boe"
        ]
      },
      "OperatorDetailDto": {
        "type": "object",
        "properties": {
          "operator_slug": {
            "type": "string",
            "example": "ypf"
          },
          "operator_name": {
            "type": "string",
            "example": "YPF S.A."
          },
          "aliases": {
            "example": [
              "YPF S.A."
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "latest_month": {
            "type": "object",
            "example": "2026-04-01",
            "nullable": true
          },
          "latest_month_rank": {
            "type": "object",
            "example": 1,
            "nullable": true
          },
          "latest": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/OperatorLatestDto"
              }
            ]
          },
          "ytd": {
            "$ref": "#/components/schemas/OperatorYtdDto"
          }
        },
        "required": [
          "operator_slug",
          "operator_name",
          "aliases",
          "latest_month",
          "latest_month_rank",
          "latest",
          "ytd"
        ]
      },
      "ApiErrorPayloadDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "NOT_FOUND"
          },
          "message": {
            "type": "string",
            "example": "Well not found: 99999"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorPayloadDto"
          }
        },
        "required": [
          "error"
        ]
      },
      "OperatorTimeSeriesPointDto": {
        "type": "object",
        "properties": {
          "date_month": {
            "type": "string",
            "example": "2026-01-01"
          },
          "oil_m3": {
            "type": "number",
            "example": 1927666.93
          },
          "oil_bbl_d": {
            "type": "number",
            "example": 404155.15
          },
          "gas_thousand_m3": {
            "type": "number",
            "example": 836299.35
          },
          "gas_mmcf_d": {
            "type": "number",
            "example": 984.46
          },
          "boe": {
            "type": "number",
            "example": 42633610.05
          },
          "active_wells": {
            "type": "number",
            "example": 5087
          }
        },
        "required": [
          "date_month",
          "oil_m3",
          "oil_bbl_d",
          "gas_thousand_m3",
          "gas_mmcf_d",
          "boe",
          "active_wells"
        ]
      },
...
      "CommoditySeriesPointDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2025-06"
          },
          "value": {
            "type": "number",
            "example": 3890
          }
        },
        "required": [
          "date",
          "value"
        ]
      },
...
      "FuelPriceRowDto": {
        "type": "object",
        "properties": {
          "provincia": {
            "type": "string",
            "example": "NEUQUEN"
          },
          "localidad": {
            "type": "object",
            "example": "NEUQUEN",
            "nullable": true
          },
          "producto": {
            "type": "string",
            "example": "Nafta Premium"
          },
          "empresa_bandera": {
            "type": "object",
            "example": "YPF",
            "nullable": true
          },
          "empresa": {
            "type": "object",
            "example": "10 DE SETIEMBRE S.A.",
            "nullable": true
          },
          "precio": {
            "type": "number",
            "example": 1250
          },
          "fecha_vigencia": {
            "type": "string",
            "example": "2026-05-01"
          },
          "tipo_horario": {
            "type": "object",
            "example": "Diurno",
            "nullable": true
          },
          "latitude": {
            "type": "object",
            "example": -38.95,
            "nullable": true
          },
          "longitude": {
            "type": "object",
            "example": -68.06,
            "nullable": true
          }
        },
        "required": [
          "provincia",
          "localidad",
          "producto",
          "empresa_bandera",
          "empresa",
          "precio",
          "fecha_vigencia",
          "tipo_horario",
          "latitude",
          "longitude"
        ]
      },
      "FuelLatestEntryDto": {
        "type": "object",
        "properties": {
          "provincia": {
            "type": "string",
            "example": "NEUQUEN"
          },
          "producto": {
            "type": "string",
            "example": "Nafta Premium"
          },
          "precio_promedio": {
            "type": "number",
            "example": 1250
          },
          "precio_min": {
            "type": "number",
            "example": 1190
          },
          "precio_max": {
            "type": "number",
            "example": 1310
          },
          "sample_size": {
            "type": "number",
            "example": 142
          },
          "fecha_vigencia": {
            "type": "string",
            "example": "2026-05-29"
          }
        },
        "required": [
          "provincia",
          "producto",
          "precio_promedio",
          "precio_min",
          "precio_max",
          "sample_size",
          "fecha_vigencia"
        ]
      },
...
    }
  }
}